Files moved and reorganized from: doc\UndergradStudents\Gordon\SimulinkLogicSimplifier_GordonTemp\src
This is a new implementation of many part of the logic simplifier to allow for simplification of subsets of systems as well as to reduce the need to handle many things through specific cases (expressions now involve the input ports which makes numerous things work more smoothly).

Support for certain blocks still needs to be added back in and subsystem_rule will not work as desired in all cases right now. 

git-svn-id: https://www.mcscert.ca/leap/svn/trunk/sw/tools/Simulink/forFCA/McMaster_Tools/SimulinkLogicSimplifier@11548 9a9db257-bd8a-4a57-83e7-af9ab8ab1356